- Dynamic Props Generater Added - Remove Unused Classes - News Feed Added - Video Manager Removed - FFmpeg Library reduced - Apk size reduced

+++ iyan3d/trunk/Iyan3D-Android/app/src/main/java/com/smackall/animator/Adapters/NoSpacingItemDecoreation.java
@@ -14,9 +14,9 @@ public class NoSpacingItemDecoreation extends RecyclerView.ItemDecoration {
     @Override
     public void getItemOffsets(Rect outRect, View view,
                                RecyclerView parent, RecyclerView.State state) {
-        outRect.bottom = (int) (Constants.height / 20);
+        outRect.bottom = Constants.height / 20;
         outRect.left = 0;
-        outRect.right =0;
+        outRect.right = 0;
         outRect.top = Constants.height / 20;
     }
 }
